requirements.txt
flask==2.3.3
flask-cors==4.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.2
accelerate==0.25.0
pymupdf==1.23.7
torch==2.1.2
nltk==3.8.1
ctransformers==0.2.27; platform_system != "Darwin" or platform_machine != "arm64"
